requirements.txt
<<<<<<< HEAD
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
asgiref==3.7.2
attrs==23.2.0
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
dnspython==2.6.1
flatbuffers==24.3.7
frozenlist==1.4.1
idna==3.6
joblib==1.3.2
multidict==6.0.5
idna==3.6
joblib==1.3.2
numpy==1.26.4
openmeteo_requests==1.2.0
openmeteo_sdk==1.10.0
pandas==2.2.1
platformdirs==4.2.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
requests-cache==1.2.0
retry-requests==2.0.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.3.0
twilio==9.0.0
typing_extensions==4.10.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
yarl==1.9.4
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
=======
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
asgiref==3.7.2
attrs==23.2.0
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
dnspython==2.6.1
flatbuffers==24.3.7
frozenlist==1.4.1
idna==3.6
joblib==1.3.2
multidict==6.0.5
idna==3.6
joblib==1.3.2
numpy==1.26.4
openmeteo_requests==1.2.0
openmeteo_sdk==1.10.0
pandas==2.2.1
platformdirs==4.2.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
requests-cache==1.2.0
retry-requests==2.0.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.3.0
twilio==9.0.0
typing_extensions==4.10.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
yarl==1.9.4
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
>>>>>>> 7ef29c090b6300e9d7ec70a747fd81796baea790
